v0.1.7: fix Windows MSVC compile in v0.1.6 SQL dialect dropdown
v0.1.6 introduced the SQL Formatter dialect dropdown which calls
SendScintilla(SCI_SETKEYWORDS, …, const char*) to inject dialect-
specific keyword sets. The call compiled fine on Linux GCC and macOS
clang but failed Windows MSVC with:

  error C2666: 'QsciScintillaBase::SendScintilla': overloaded
  functions have similar conversions

Root cause: ambiguity between two SendScintilla overloads:
  - SendScintilla(unsigned int, unsigned long, void *)
  - SendScintilla(unsigned int, uintptr_t, const char *)

On MSVC x64 (LLP64), unsigned long is 32-bit but uintptr_t is 64-bit,
so neither overload was strictly better — the first matched wParam
exactly but needed const_cast on lParam, the second matched lParam
exactly but needed an integer widening on wParam. Compiler picked
neither and emitted C2666.

On Linux x64 (LP64), unsigned long IS uintptr_t (both 64-bit), so the
second overload was strictly better and GCC picked it cleanly.

Fix: cast wParam to (uintptr_t)0 explicitly so the (uintptr_t,
const char *) overload is an exact match on both compilers.

v0.1.6 built green on Linux + macOS but the Windows job failed at
the C++ compile step before any artifact upload, so no GitHub Release
was published. v0.1.7 ships everything from v0.1.6 with the MSVC
fix — same plugin overhaul, same SQL dialect feature, same lighter
font, same compare-unsaved-tabs improvement, just with a Windows
binary that actually compiles.

Bumped via scripts/bump_version.sh (CMakeLists.txt project version,
docs/index.html, README.md, CHANGELOG.md, release_notes/v0.1.7.md).

+++ src/sqlfmtpanel.cpp
@@ -227,8 +227,14 @@ void SqlFmtPanel::applySqlDialectKeywords() {
 
     // QsciLexerSQL::setKeywords is protected — go straight to Scintilla
     // via SCI_SETKEYWORDS, which works for any lexer.
+    //
+    // Cast wParam to uintptr_t (not unsigned long) so MSVC unambiguously
+    // picks the SendScintilla(uint, uintptr_t, const char*) overload.
+    // On MSVC x64, unsigned long is 32-bit and uintptr_t is 64-bit, so a
+    // bare 0 + const char* is ambiguous between (unsigned long, void*) and
+    // (uintptr_t, const char*). Casting to uintptr_t makes the latter exact.
     m_output->SendScintilla(QsciScintilla::SCI_SETKEYWORDS,
-                            (unsigned long)0,
+                            (uintptr_t)0,
                             keywords.constData());
     // Re-colourise the visible buffer so the new keywords paint immediately
     int len = m_output->length();
